requirements.txt
Pillow==10.0.0
python-tk==0.1.0